Final_project/requirements.txt
fastapi==0.115.0
uvicorn==0.32.1
python-multipart==0.0.20
jinja2==3.1.4
torch==2.5.1
torchvision==0.20.1
joblib==1.4.2
pandas==2.2.3
numpy==2.2.1
pillow==11.0.0
opencv-python==4.10.0.84
scikit-learn==1.6.0
